chore: fix testing lib
