+++ back-end/src/db/index.ts
@@ -1,3 +1,4 @@
+import "@/env";
 import { drizzle, type PostgresJsDatabase } from "drizzle-orm/postgres-js";
 import postgres from "postgres";
 import * as schema from "./schema";
