+++ back-end/src/test/setup.ts
@@ -1,0 +1,4 @@
+import "@/env";
+
+// This file is run before all tests to ensure environment variables are loaded
+// The env module loads .env.test when NODE_ENV=test
