+++ back-end/src/users/actions.ts
@@ -5,11 +5,15 @@ import { UserAlreadyOnboardedError, UserAlreadyVerifiedError } from "./types";
 
 export const createUser = (params: CreateUserParams): Result<User, never> => {
   const uuid = generateUuid();
+  const now = new Date();
   return ok({
     id: uuid,
     ...params,
     emailVerified: false,
     onboarded: false,
+    image: null,
+    createdAt: now,
+    updatedAt: now,
   });
 };
