+++ back-end/src/users/types.ts
@@ -1,6 +1,6 @@
 import { createError } from "@/lib/utils/createError";
 import z from "zod";
-import { createInsertSchema } from "drizzle-zod";
+import { createInsertSchema, createSelectSchema } from "drizzle-zod";
 import { users } from "@/db/schema";
 
 // --------------------------------
@@ -8,17 +8,7 @@ import { users } from "@/db/schema";
 // --------------------------------
 
 export const userInsertSchema = createInsertSchema(users);
-
-export const userSchema = z.object({
-  id: z.uuid(),
-  name: z.string(),
-  email: z.email(),
-  emailVerified: z.boolean(),
-  onboarded: z.boolean(),
-  image: z.string().nullable().optional(),
-  createdAt: z.date().optional(),
-  updatedAt: z.date().optional(),
-});
+export const userSchema = createSelectSchema(users);
 
 export type User = z.infer<typeof userSchema>;
 
@@ -63,18 +53,7 @@ export const UserAlreadyExistsError = createError(
   },
 );
 
-export const UserEmailAlreadyInUseError = createError(
-  "UserEmailAlreadyInUseError",
-  (email: string) => `User with email: ${email} already exists`,
-  {
-    code: "USER_EMAIL_ALREADY_IN_USE",
-    error: "Validation Error",
-    statusCode: 422,
-  },
-);
-
 export type UserValidationError =
   | InstanceType<typeof UserAlreadyOnboardedError>
   | InstanceType<typeof UserAlreadyExistsError>
-  | InstanceType<typeof UserAlreadyVerifiedError>
-  | InstanceType<typeof UserEmailAlreadyInUseError>;
+  | InstanceType<typeof UserAlreadyVerifiedError>;
