+++ back-end/vitest.config.ts
@@ -3,4 +3,10 @@ import tsconfigPaths from "vite-tsconfig-paths";
 
 export default defineConfig({
   plugins: [tsconfigPaths()],
+  test: {
+    setupFiles: ["./src/test/setup.ts"],
+    env: {
+      NODE_ENV: "test",
+    },
+  },
 });
